Replace OctahedralProjectedCuCubeMap with SpecularEnvironmentCubeMap
