+++ packages/engine/Source/Scene/Model/ImageBasedLightingPipelineStage.js
@@ -2,7 +2,7 @@ import combine from "../../Core/combine.js";
 import defined from "../../Core/defined.js";
 import ImageBasedLightingStageFS from "../../Shaders/Model/ImageBasedLightingStageFS.js";
 import ShaderDestination from "../../Renderer/ShaderDestination.js";
-import OctahedralProjectedCubeMap from "../OctahedralProjectedCubeMap.js";
+import SpecularEnvironmentCubeMap from "../SpecularEnvironmentCubeMap.js";
 
 const ImageBasedLightingPipelineStage = {
   name: "ImageBasedLightingPipelineStage", // Helps with debugging
@@ -34,7 +34,7 @@ ImageBasedLightingPipelineStage.process = function (
     ShaderDestination.FRAGMENT
   );
 
-  if (OctahedralProjectedCubeMap.isSupported(frameState.context)) {
+  if (SpecularEnvironmentCubeMap.isSupported(frameState.context)) {
     const addMatrix =
       imageBasedLighting.useSphericalHarmonics ||
       imageBasedLighting.useSpecularEnvironmentMaps ||
@@ -86,13 +86,8 @@ ImageBasedLightingPipelineStage.process = function (
         ShaderDestination.FRAGMENT
       );
       shaderBuilder.addUniform(
-        "sampler2D",
+        "samplerCube",
         "model_specularEnvironmentMaps",
-        ShaderDestination.FRAGMENT
-      );
-      shaderBuilder.addUniform(
-        "vec2",
-        "model_specularEnvironmentMapsSize",
         ShaderDestination.FRAGMENT
       );
       shaderBuilder.addUniform(
@@ -139,9 +134,6 @@ ImageBasedLightingPipelineStage.process = function (
     },
     model_specularEnvironmentMaps: function () {
       return imageBasedLighting.specularEnvironmentMapAtlas.texture;
-    },
-    model_specularEnvironmentMapsSize: function () {
-      return imageBasedLighting.specularEnvironmentMapAtlas.texture.dimensions;
     },
     model_specularEnvironmentMapsMaximumLOD: function () {
       return imageBasedLighting.specularEnvironmentMapAtlas.maximumMipmapLevel;
