+++ packages/engine/Source/Scene/SkyBox.js
@@ -1,6 +1,7 @@
 import buildModuleUrl from "../Core/buildModuleUrl.js";
 import BoxGeometry from "../Core/BoxGeometry.js";
 import Cartesian3 from "../Core/Cartesian3.js";
+import Check from "../Core/Check.js";
 import defaultValue from "../Core/defaultValue.js";
 import defined from "../Core/defined.js";
 import destroyObject from "../Core/destroyObject.js";
@@ -95,49 +96,38 @@ function SkyBox(options) {
  */
 SkyBox.prototype.update = function (frameState, useHdr) {
   const that = this;
+  const { mode, passes, context } = frameState;
 
   if (!this.show) {
     return undefined;
   }
 
-  if (
-    frameState.mode !== SceneMode.SCENE3D &&
-    frameState.mode !== SceneMode.MORPHING
-  ) {
+  if (mode !== SceneMode.SCENE3D && mode !== SceneMode.MORPHING) {
     return undefined;
   }
 
   // The sky box is only rendered during the render pass; it is not pickable, it doesn't cast shadows, etc.
-  if (!frameState.passes.render) {
+  if (!passes.render) {
     return undefined;
   }
-
-  const context = frameState.context;
 
   if (this._sources !== this.sources) {
     this._sources = this.sources;
     const sources = this.sources;
 
     //>>includeStart('debug', pragmas.debug);
-    if (
-      !defined(sources.positiveX) ||
-      !defined(sources.negativeX) ||
-      !defined(sources.positiveY) ||
-      !defined(sources.negativeY) ||
-      !defined(sources.positiveZ) ||
-      !defined(sources.negativeZ)
-    ) {
+    Check.defined("this.sources", sources);
+    if (CubeMap.faceNames.some((faceName) => !defined(sources[faceName]))) {
       throw new DeveloperError(
-        "this.sources is required and must have positiveX, negativeX, positiveY, negativeY, positiveZ, and negativeZ properties."
+        "this.sources must have positiveX, negativeX, positiveY, negativeY, positiveZ, and negativeZ properties."
       );
     }
 
+    const sourceType = typeof sources.positiveX;
     if (
-      typeof sources.positiveX !== typeof sources.negativeX ||
-      typeof sources.positiveX !== typeof sources.positiveY ||
-      typeof sources.positiveX !== typeof sources.negativeY ||
-      typeof sources.positiveX !== typeof sources.positiveZ ||
-      typeof sources.positiveX !== typeof sources.negativeZ
+      CubeMap.faceNames.some(
+        (faceName) => typeof sources[faceName] !== sourceType
+      )
     ) {
       throw new DeveloperError(
         "this.sources properties must all be the same type."
